production rate ui tweak and auto-sized-input refactor
