+++ src/natural_production_line.jsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import { GlobalStateContext, GameInfoContext, UiSettingsContext, UiSettingsSetterContext } from './contexts.jsx';
 import { FactorySelect, ProModeSelect, ProNumSelect, RecipeSelect } from './result.jsx';
 import { ItemSelect } from './item_select.jsx';
-import { AutoSizedInput } from './auto_sized_input.jsx';
+import { AutoSizedInput } from './ui_components/auto_sized_input.jsx';
 import { ItemIcon } from './icon';
 
 // { "目标物品": "氢", "建筑数量": 0, "配方id": 1, "喷涂点数": 0, "增产模式": 0, "建筑": 0 }
