+++ src/ui_components/auto_sized_input.jsx
@@ -1,0 +1,8 @@
+export const AutoSizedInput = ({ value, onChange, className }) => {
+  return (
+    <label className={`auto-sized-input ${className || ""}`}>
+      <span>{value}</span>
+      <input type="text" value={value} onChange={onChange} />
+    </label>
+  )
+}
